Fix elf/tst-ldconfig-X for cross testing.
elf/Makefile passes arguments to tst-ldconfig-X.sh that are different
from what it expects, so resulting in the test failing in cross
testing.  This patch corrects the arguments passed (the script itself
has correct logic for cross testing, it's just the Makefile that's
wrong).

Tested for powerpc (cross testing) and for x86_64 (native testing).

	* elf/Makefile ($(objpfx)tst-ldconfig-X.out): Correct arguments
	passed to tst-ldconfig-X.sh.