Move most libmvec test contents from .c to .h files.
The libmvec tests put substantive, architecture-specific contents in
.c files such as test-double-vlen4.c, so making those files
architecture-specific and causing issues for generating such files
automatically when splitting up tests by function.

This patch moves all the substantive contents to .h files, so the .c
files only include the .h file and then libm-test.c.  This allows for
automatic generation of per-function .c files in future.  The .h files
in turn #include or #include_next the architecture-independent file
and add the architecture-specific definitions to that.  (Splitting by
function should in fact allow the TEST_VECTOR_* macros to be replaced
by sysdeps makefile information on which functions to test in each
case, removing the need for gen-libm-have-vector-test.sh as well as
removing the need for some of the architecture-specific headers.)

Tested for x86_64.

	* sysdeps/x86_64/fpu/test-double-vlen2.c: Move most contents to,
	and include ...
	* sysdeps/x86_64/fpu/test-double-vlen2.h: ... here.  New file.
	* sysdeps/x86_64/fpu/test-double-vlen4-avx2.c: Move most contents
	to, and include ...
	* sysdeps/x86_64/fpu/test-double-vlen4-avx2.h: ... here.  New
	file.
	* sysdeps/x86_64/fpu/test-double-vlen4.c: Move most contents to,
	and include ...
	* sysdeps/x86_64/fpu/test-double-vlen4.h: ... here.  New file.
	* sysdeps/x86_64/fpu/test-double-vlen8.c: Move most contents to,
	and include ...
	* sysdeps/x86_64/fpu/test-double-vlen8.h: ... here.  New file.
	* sysdeps/x86_64/fpu/test-float-vlen16.c: Move most contents to,
	and include ...
	* sysdeps/x86_64/fpu/test-float-vlen16.h: ... here.  New file.
	* sysdeps/x86_64/fpu/test-float-vlen4.c: Move most contents to,
	and include ...
	* sysdeps/x86_64/fpu/test-float-vlen4.h: ... here.  New file.
	* sysdeps/x86_64/fpu/test-float-vlen8-avx2.c: Move most contents
	to, and include ...
	* sysdeps/x86_64/fpu/test-float-vlen8-avx2.h: ... here.  New file.
	* sysdeps/x86_64/fpu/test-float-vlen8.c: Move most contents to,
	and include ...
	* sysdeps/x86_64/fpu/test-float-vlen8.h: ... here.  New file.